+++ spring-boot-bookstore/catalog-service/src/main/java/com/bookstore/catalogservice/product/command/api/commands/CreateProductCommand.java
@@ -5,7 +5,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
+
 
 @Data
 @Getter
@@ -18,7 +18,7 @@ public class CreateProductCommand {
 
     private String description;
 
-    private BigDecimal price;
+    private double price;
 
     private int availableItemCount;
