+++ spring-boot-bookstore/catalog-service/src/main/java/com/bookstore/catalogservice/product/command/write_model/request/CreateProductRequest.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 
-import java.math.BigDecimal;
+
 
 @Data
 @Getter
@@ -16,7 +16,7 @@ public class CreateProductRequest {
 
     private String description;
 
-    private BigDecimal price;
+    private double price;
 
     private int availableItemCount;
