+++ spring-boot-bookstore/catalog-service/src/main/java/com/bookstore/catalogservice/product/command/write_model/request/UpdateProductRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
+
 
 
 @Data
@@ -19,7 +19,7 @@ public class UpdateProductRequest {
 
     private String description;
 
-    private BigDecimal price;
+    private double price;
 
     private int availableItemCount;
